# dispatcher_microservice/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a dispatcher microservice for Factory Games Organizer api. Api is responsible for creating, updating and authenicating api users, CRUD operations on database associated with the api and provides production calculator service.",
        "title": "Dispatcher microservice",
        "contact": {
            "name": "Marek Banaś",
            "email": "marek.banas004@gmail.com"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "1.0"
    },
    "host": "79.175.222.18:40000",
    "basePath": "/",
    "paths": {
        "/calculator/calculate": {
            "get": {
                "description": "Calculate the machines and resources needed to produce target resource with provided production rate per second. Alternative Recipe and Alternative Machine parameters can be present multiple times in request query.",
                "tags": [
                    "Calculator"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of users whose data will be used as the base for calculation",
                        "name": "userid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource to be produced",
                        "name": "resource",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target production rate for the specified resource",
                        "name": "rate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Alternative recipe to take into consideration when calculating production tree",
                        "name": "alt_recipe",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Alternative machine to take into consideration when calculating production tree",
                        "name": "alt_machine",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ProductionTreeCalculator"
                        }
                    },
                    "400": {
                        "description": "Bad request. One of required parameters is missing",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Unexpected serverside error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/stats": {
            "get": {
                "description": "Return the usage stats of microservice.",
                "tags": [
                    "Dispatcher"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.StatsResponse"
                        }
                    },
                    "500": {
                        "description": "Unexpected serverside error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.ProductionTreeCalculator": {
            "type": "object",
            "properties": {
                "excessResources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.ResourceSource"
                    }
                },
                "targetResource": {
                    "type": "string"
                },
                "targetResourceRate": {
                    "type": "number",
                    "format": "float32"
                },
                "targetResourceSourceNode": {
                    "type": "integer"
                },
                "treeNodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.ProductionTreeNode"
                    }
                }
            }
        },
        "handler.ProductionTreeNode": {
            "type": "object",
            "properties": {
                "machineName": {
                    "type": "string"
                },
                "machineNumber": {
                    "type": "number",
                    "format": "float32"
                },
                "nodeId": {
                    "type": "integer"
                },
                "producedResourcesPerSecond": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float32"
                    }
                },
                "recipeName": {
                    "type": "string"
                },
                "requiredResourcesPerSecond": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float32"
                    }
                },
                "sourceNodes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "totalPowerConsumedkW": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
        "handler.ResourceSource": {
            "type": "object",
            "properties": {
                "excessProducedResourcePerSecond": {
                    "type": "number",
                    "format": "float32"
                },
                "excessResourceName": {
                    "type": "string"
                },
                "nodeId": {
                    "type": "integer"
                }
            }
        },
        "handler.StatsResponse": {
            "type": "object",
            "properties": {
                "apiUsageStats": {
                    "type": "object"
                },
                "noPeriods": {
                    "type": "integer",
                    "format": "int64"
                },
                "trackingPeriodMs": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "apiTokenAuth": {
            "type": "apiKey",
            "name": "jwt",
            "in": "query"
        }
    }
}